Creating Kernel Modules
After performing the service syslog restart command, I saw the Shutting down kernel and system logger and the Starting system and kernel logger process successfully complete in that order. In the var/log/messages file I see how the kernel configures and connects with the hardware devices and learnt more on how exactly software communicates messages to the hardware. The command lsmod shows the modules loaded onto the kernel at the moment so when I unload my module with the rmmod simple command then simple is removed from lsmod’s results, the content of the modules folder.